requirements.txt
requests
openmeteo-requests
pandas
requests-cache
retry-requests
snowflake-connector-python
dotenv